mobile-backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
sqlalchemy==2.0.23
pymysql==1.1.0
python-multipart==0.0.6
qrcode[pil]==7.4.2
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4